+++ lib/auth.php
@@ -2,7 +2,7 @@
 
 /*
 
-	Copyright (c) 2009-2018 F3::Factory/Bong Cosca, All rights reserved.
+	Copyright (c) 2009-2019 F3::Factory/Bong Cosca, All rights reserved.
 
 	This file is part of the Fat-Free Framework (http://fatfreeframework.com).