+++ lib/bcrypt.php
@@ -1,7 +1,7 @@
 <?php
 
 /**
-*	Copyright (c) 2009-2017 F3::Factory/Bong Cosca, All rights reserved.
+*	Copyright (c) 2009-2019 F3::Factory/Bong Cosca, All rights reserved.
 *
 *	This file is part of the Fat-Free Framework (http://fatfreeframework.com).
 *