+++ lib/cli/ws.php
@@ -2,7 +2,7 @@
 
 /*
 
-	Copyright (c) 2009-2017 F3::Factory/Bong Cosca, All rights reserved.
+	Copyright (c) 2009-2019 F3::Factory/Bong Cosca, All rights reserved.
 
 	This file is part of the Fat-Free Framework (http://fatfreeframework.com).
 
@@ -138,11 +138,13 @@ class WS {
 	*	Read from stream socket
 	*	@return string|FALSE
 	*	@param $socket resource
+	*	@param $len int
 	**/
-	function read($socket) {
-		if (is_string($buf=@fread($socket,WS::Packet)) &&
-			strlen($buf) &&
-			strlen($buf)<WS::Packet)
+	function read($socket,$len=0) {
+		if (!$len)
+			$len=WS::Packet;
+		if (is_string($buf=@fread($socket,$len)) &&
+			strlen($buf) && strlen($buf)<$len)
 			return $buf;
 		if (isset($this->events['error']) &&
 			is_callable($func=$this->events['error']))
@@ -179,6 +181,10 @@ class WS {
 	function agents($uri=NULL) {
 		return array_filter(
 			$this->agents,
+			/**
+			 * @var $val Agent
+			 * @return bool
+			 */
 			function($val) use($uri) {
 				return $uri?($val->uri()==$uri):TRUE;
 			}
@@ -206,19 +212,15 @@ class WS {
 
 	/**
 	*	Terminate server
-	*	@return NULL
-	*	@param $signal int
 	**/
-	function kill($signal) {
+	function kill() {
 		die;
 	}
 
 	/**
 	*	Execute the server process
-	*	@return object
 	**/
 	function run() {
-		$fw=\Base::instance();
 		// Assign signal handlers
 		declare(ticks=1);
 		pcntl_signal(SIGINT,[$this,'kill']);
@@ -303,8 +305,6 @@ class WS {
 	}
 
 	/**
-	*	Instantiate object
-	*	@return object
 	*	@param $addr string
 	*	@param $ctx resource
 	*	@param $wait int
@@ -333,7 +333,7 @@ class Agent {
 
 	/**
 	*	Return server instance
-	*	@return object
+	*	@return WS
 	**/
 	function server() {
 		return $this->server;
@@ -349,7 +349,7 @@ class Agent {
 
 	/**
 	*	Return socket
-	*	@return object
+	*	@return resource
 	**/
 	function socket() {
 		return $this->socket;
@@ -373,7 +373,7 @@ class Agent {
 
 	/**
 	*	Return socket headers
-	*	@return string
+	*	@return array
 	**/
 	function headers() {
 		return $this->headers;
@@ -382,9 +382,8 @@ class Agent {
 	/**
 	*	Frame and transmit payload
 	*	@return string|FALSE
-	*	@param $socket resource
 	*	@param $op int
-	*	@param $payload string
+	*	@param $data string
 	**/
 	function send($op,$data='') {
 		$server=$this->server;
@@ -393,8 +392,7 @@ class Agent {
 		$buf='';
 		if ($len>0xffff)
 			$buf=pack('CCNN',$mask,0x7f,$len);
-		else
-		if ($len>0x7d)
+		elseif ($len>0x7d)
 			$buf=pack('CCn',$mask,0x7e,$len);
 		else
 			$buf=pack('CC',$mask,$len);
@@ -410,55 +408,56 @@ class Agent {
 
 	/**
 	*	Retrieve and unmask payload
-	*	@return array|FALSE
+	*	@return bool|NULL
 	**/
 	function fetch() {
 		// Unmask payload
 		$server=$this->server;
 		if (is_bool($buf=$server->read($this->socket)))
 			return FALSE;
-		$op=ord($buf[0]) & WS::OpCode;
-		$len=ord($buf[1]) & WS::Length;
-		$pos=2;
-		if ($len==0x7e) {
-			$len=ord($buf[2])*256+ord($buf[3]);
-			$pos+=2;
+		while($buf) {
+			$op=ord($buf[0]) & WS::OpCode;
+			$len=ord($buf[1]) & WS::Length;
+			$pos=2;
+			if ($len==0x7e) {
+				$len=ord($buf[2])*256+ord($buf[3]);
+				$pos+=2;
+			}
+			else
+			if ($len==0x7f) {
+				for ($i=0,$len=0;$i<8;$i++)
+					$len=$len*256+ord($buf[$i+2]);
+				$pos+=8;
+			}
+			for ($i=0,$mask=[];$i<4;$i++)
+				$mask[$i]=ord($buf[$pos+$i]);
+			$pos+=4;
+			if (strlen($buf)<$len+$pos)
+				return FALSE;
+			for ($i=0,$data='';$i<$len;$i++)
+				$data.=chr(ord($buf[$pos+$i])^$mask[$i%4]);
+			// Dispatch
+			switch ($op & WS::OpCode) {
+			case WS::Ping:
+				$this->send(WS::Pong);
+				break;
+			case WS::Close:
+				$server->close($this->socket);
+				break;
+			case WS::Text:
+				$data=trim($data);
+			case WS::Binary:
+				if (isset($this->events['receive']) &&
+					is_callable($func=$this->events['receive']))
+					$func($this,$op,$data);
+				break;
+			}
+			$buf = substr($buf, $len+$pos);
 		}
-		else
-		if ($len==0x7f) {
-			for ($i=0,$len=0;$i<8;$i++)
-				$len=$len*256+ord($buf[$i+2]);
-			$pos+=8;
-		}
-		for ($i=0,$mask=[];$i<4;$i++)
-			$mask[$i]=ord($buf[$pos+$i]);
-		$pos+=4;
-		if (strlen($buf)<$len+$pos)
-			return FALSE;
-		for ($i=0,$data='';$i<$len;$i++)
-			$data.=chr(ord($buf[$pos+$i])^$mask[$i%4]);
-		// Dispatch
-		switch ($op & WS::OpCode) {
-		case WS::Ping:
-			$this->send(WS::Pong);
-			break;
-		case WS::Close:
-			$server->close($this->socket);
-			break;
-		case WS::Text:
-			$data=trim($data);
-		case WS::Binary:
-			if (isset($this->events['receive']) &&
-				is_callable($func=$this->events['receive']))
-				$func($this,$op,$data);
-			break;
-		}
-		return [$op,$data];
 	}
 
 	/**
 	*	Destroy object
-	*	@return NULL
 	**/
 	function __destruct() {
 		if (isset($this->events['disconnect']) &&
@@ -467,9 +466,7 @@ class Agent {
 	}
 
 	/**
-	*	Instantiate object
-	*	@return object
-	*	@param $server object
+	*	@param $server WS
 	*	@param $socket resource
 	*	@param $verb string
 	*	@param $uri string
@@ -483,6 +480,7 @@ class Agent {
 		$this->uri=$uri;
 		$this->headers=$hdrs;
 		$this->events=$server->events();
+
 		if (isset($this->events['connect']) &&
 			is_callable($func=$this->events['connect']))
 			$func($this);