+++ lib/web.php
@@ -2,7 +2,7 @@
 
 /*
 
-	Copyright (c) 2009-2017 F3::Factory/Bong Cosca, All rights reserved.
+	Copyright (c) 2009-2019 F3::Factory/Bong Cosca, All rights reserved.
 
 	This file is part of the Fat-Free Framework (http://fatfreeframework.com).
 
@@ -33,11 +33,87 @@ class Web extends Prefab {
 		$wrapper;
 
 	/**
-	*	Detect MIME type using file extension
+	*	Detect MIME type using file extension or file inspection
 	*	@return string
 	*	@param $file string
+	*	@param $inspect bool
 	**/
-	function mime($file) {
+	function mime($file, $inspect=FALSE) {
+		if ($inspect) {
+			if (is_file($file) && is_readable($file)) {
+				// physical files
+				if (extension_loaded('fileinfo'))
+					$mime=mime_content_type($file);
+				elseif (preg_match('/Darwin/i',PHP_OS))
+					$mime=trim(exec('file -bI '.escapeshellarg($file)));
+				elseif (!preg_match('/^win/i',PHP_OS))
+					$mime=trim(exec('file -bi '.escapeshellarg($file)));
+				if (isset($mime) && !empty($mime)){
+					// cut charset information if any
+					$exp=explode(';',$mime,2);
+					$mime=$exp[0];
+				}
+			}
+			else {
+				// remote and stream files
+				if (ini_get('allow_url_fopen') && ($fhandle=fopen($file,'rb'))) {
+					// only get head bytes instead of whole file
+					$bytes=fread($fhandle,20);
+					fclose($fhandle);
+				}
+				elseif (($response=$this->request($file,['method' => 'HEAD']))
+					&& preg_grep('/HTTP\/\d\.\d 200/',$response['headers'])
+					&& ($type = preg_grep('/^Content-Type:/i',$response['headers']))) {
+					// get mime type directly from response header
+					return preg_replace('/^Content-Type:\s*/i','',array_pop($type));
+				}
+				else // load whole file
+					$bytes=file_get_contents($file);
+				if (extension_loaded('fileinfo')) {
+					// get mime from fileinfo
+					$finfo=finfo_open(FILEINFO_MIME_TYPE);
+					$mime=finfo_buffer($finfo,$bytes);
+				}
+				elseif ($bytes) {
+					// magic number header fallback
+					$map=[
+						'\x64\x6E\x73\x2E'=>'audio/basic',
+						'\x52\x49\x46\x46.{4}\x41\x56\x49\x20\x4C\x49\x53\x54'=>'video/avi',
+						'\x42\x4d'=>'image/bmp',
+						'\x42\x5A\x68'=>'application/x-bzip2',
+						'\x07\x64\x74\x32\x64\x64\x74\x64'=>'application/xml-dtd',
+						'\xD0\xCF\x11\xE0\xA1\xB1\x1A\xE1'=>'application/msword',
+						'\x50\x4B\x03\x04\x14\x00\x06\x00'=>'application/msword',
+						'\x0D\x44\x4F\x43'=>'application/msword',
+						'GIF\d+a'=>'image/gif',
+						'\x1F\x8B'=>'application/x-gzip',
+						'\xff\xd8\xff'=>'image/jpeg',
+						'\x49\x46\x00'=>'image/jpeg',
+						'\xFF\xFB'=>'audio/mpeg',
+						'\x49\x44\x33'=>'audio/mpeg',
+						'\x00\x00\x01\xBA'=>'video/mpeg',
+						'\x4F\x67\x67\x53\x00\x02\x00\x00\x00\x00\x00\x00\x00\x00'=>'audio/vorbis',
+						'\x25\x50\x44\x46'=>'application/pdf',
+						'\x89PNG\x0d\x0a'=>'image/png',
+						'.{4}\x6D\x6F\x6F\x76\x'=>'video/quicktime',
+						'\x53\x49\x54\x21\x00'=>'application/x-stuffit',
+						'\x43\x57\x53'=>'application/x-shockwave-flash',
+						'\x1F\x8B\x08'=>'application/x-tar',
+						'\x49\x20\x49'=>'image/tiff',
+						'\x52\x49\x46\x46.{4}\x57\x41\x56\x45\x66\x6D\x74\x20'=>'audio/wav',
+						'\xFD\xFF\xFF\xFF\x20\x00\x00\x00'=>'application/vnd.ms-excel',
+						'\x50\x4B\x03\x04'=>'application/x-zip-compressed',
+						'[ -~]+$'=>'text/plain',
+					];
+					foreach ($map as $key=>$val)
+						if (preg_match('/^'.$key.'/',substr($bytes,0,128)))
+							return $val;
+				}
+			}
+			if (isset($mime) && !empty($mime))
+				return $mime;
+			// Fallback to file extension-based check if no mime was found yet
+		}
 		if (preg_match('/\w+$/',$file,$ext)) {
 			$map=[
 				'au'=>'audio/basic',
@@ -617,7 +693,7 @@ class Web extends Prefab {
 		$dst='';
 		if (!isset($path))
 			$path=$fw->UI.';./';
-		foreach ($fw->split($path,FALSE) as $dir)
+		foreach (array_unique($fw->split($path,FALSE)) as $dir)
 			foreach ($files as $file)
 				if (is_file($save=$fw->fixslashes($dir.$file)) &&
 					is_bool(strpos($save,'../')) &&
@@ -639,6 +715,12 @@ class Web extends Prefab {
 									($path?($path.'/'):'').$parts[2],
 									$mime,$header
 								);
+								$ptr+=strlen($parts[0]);
+								continue;
+							}
+							if ($ext[0]=='css'&&preg_match('/^url\(([^\'"].*?[^\'"])\)/i',
+									substr($src,$ptr),$parts)) {
+								$data.=$parts[0];
 								$ptr+=strlen($parts[0]);
 								continue;
 							}
@@ -733,6 +815,7 @@ class Web extends Prefab {
 							$cache->set($hash,$data);
 						$dst.=$data;
 					}
+					break;
 				}
 		if (PHP_SAPI!='cli' && $header)
 			header('Content-Type: '.$mime.'; charset='.$fw->ENCODING);
@@ -855,6 +938,7 @@ class Web extends Prefab {
 			'ǜ'=>'u','ǔ'=>'u','ǖ'=>'u','ũ'=>'u','ü'=>'ue','в'=>'v',
 			'ŵ'=>'w','ы'=>'y','ÿ'=>'y','ý'=>'y','ŷ'=>'y','ź'=>'z',
 			'ž'=>'z','з'=>'z','ż'=>'z','ж'=>'zh','ь'=>'','ъ'=>'',
+			'њ'=>'nj','љ'=>'lj','ђ'=>'dj','џ'=>'dz','ћ'=>'c','ј'=>'j',
 			'\''=>'',
 		];
 	}
@@ -915,6 +999,7 @@ if (!function_exists('gzdecode')) {
 	/**
 	*	Decode gzip-compressed string
 	*	@param $str string
+	*	@return string
 	**/
 	function gzdecode($str) {
 		$fw=Base::instance();